isolating the keymap normalizer logic in a single module
