+++ lib/utils/keymaps.js
@@ -1,24 +1,7 @@
 import {remote} from 'electron';
+import {returnCommand} from './keymaps-normalize';
 
 const {getKeymaps} = remote.require('./config');
-
-function normalize(keybinding) {
-  function sortAlphabetically(a, b) {
-    return a.localeCompare(b);
-  }
-  return keybinding.split('+').sort(sortAlphabetically).join('+');
-}
-
-function getNormalizedKeyMaps() {
-  const userKeymaps = getKeymaps().keys;
-  const normalized = {};
-  for (const keymap in userKeymaps) {
-    if (userKeymaps[keymap]) {
-      normalized[normalize(keymap)] = userKeymaps[keymap];
-    }
-  }
-  return normalized;
-}
 
 export default function returnKey(e) {
   let keys = [];
@@ -47,6 +30,6 @@ export default function returnKey(e) {
     keys.push(e.key.replace('Arrow', ''));
   }
 
-  keys = normalize(keys.join('+'));
-  return getNormalizedKeyMaps()[keys.toLowerCase()];
+  keys = keys.join('+');
+  return returnCommand(keys, getKeymaps().keys);
 }
